Phonemes2Sound Module uploaded
Uploaded a test version of the Phonemes2Sound module. This module takes as an input one phonemes string (or many of ones) and generates the audio file with the help of Amazon AWS. The main target of the module is to provide the possibility for users to listen words they learn in cases when FreeDictionary doesn't provice such possibility.